// frontend/web/src/services/API/proserve-wb-provisioning-api/models/GetAvailableProductsResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Product Provisioning API
 * API in charge of provisioning VEW products
 *
 * The version of the OpenAPI document: 2024-08-12T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { AvailableProduct } from './AvailableProduct';
import {
    AvailableProductFromJSON,
    AvailableProductFromJSONTyped,
    AvailableProductToJSON,
} from './AvailableProduct';

/**
 * 
 * @export
 * @interface GetAvailableProductsResponse
 */
export interface GetAvailableProductsResponse {
    /**
     * 
     * @type {Array<AvailableProduct>}
     * @memberof GetAvailableProductsResponse
     */
    availableProducts: Array<AvailableProduct>;
}

/**
 * Check if a given object implements the GetAvailableProductsResponse interface.
 */
export function instanceOfGetAvailableProductsResponse(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "availableProducts" in value;

    return isInstance;
}

export function GetAvailableProductsResponseFromJSON(json: any): GetAvailableProductsResponse {
    return GetAvailableProductsResponseFromJSONTyped(json, false);
}

export function GetAvailableProductsResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): GetAvailableProductsResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'availableProducts': ((json['availableProducts'] as Array<any>).map(AvailableProductFromJSON)),
    };
}

export function GetAvailableProductsResponseToJSON(value?: GetAvailableProductsResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'availableProducts': ((value.availableProducts as Array<any>).map(AvailableProductToJSON)),
    };
}